Refactored wpandas
