+++ wpandas/core/groupby/__init__.py
@@ -1,0 +1,1 @@
+from wpandas.core.groupby.generic import WeightedSeriesGroupBy, WeightedDataFrameGroupBy
