+++ wpandas/core/util/random.py
@@ -1,0 +1,14 @@
+"""Random number generation utilities."""
+import contextlib
+import numpy as np
+
+
+@contextlib.contextmanager
+def temporary_seed(seed):
+    """Context for temporarily setting a numpy seed."""
+    state = np.random.get_state()
+    np.random.seed(seed)
+    try:
+        yield
+    finally:
+        np.random.set_state(state)
